QA: fix return type of _get_library_paths
The tuple is not of a fixed length. We also have to explicitly declare
the type of the variable in the function, otherwise mypy infers it from
the first declaration (os_is_nt) where it is a one-element tuple,
causing mismatches in the other branches of the "if" statement.
See https://stackoverflow.com/a/52860650 and
https://mypy.readthedocs.io/en/stable/type_inference_and_annotations.html.

(cherry picked from commit 2a4650ece0534b68a22e01c34f9bc179001a2ec4)